+++ package.json
@@ -10,12 +10,17 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.11",
+    "@react-navigation/native": "^7.0.0",
+    "@react-navigation/stack": "^7.0.0",
     "chess.js": "^1.0.0-beta.3",
     "react": "18.2.0",
     "react-native": "0.71.3",
-    "react-native-gesture-handler": "^2.9.0",
-    "react-native-reanimated": "^2.14.4",
-    "react-native-redash": "^18.1.0"
+    "react-native-gesture-handler": "^2.20.2",
+    "react-native-reanimated": "^2.17.0",
+    "react-native-redash": "^18.1.0",
+    "react-native-safe-area-context": "^4.14.0",
+    "react-native-screens": "^3.35.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
